// sources/client/app/actions/actionsTypes.ts
export interface Action {
  type: ActionType;
  payload: any;
}

export enum ActionType {
  UPDATE_SETTINGS = 'UPDATE_SETTINGS',
  UPDATE_MODULES_CONFIG = 'UPDATE_MODULES_CONFIG',
  SET_LOCALES = 'SET_LOCALES',
  SET_LOCALE = 'SET_LOCALE',
  SET_INTERNET_CONNECTION = 'INTERNET_CONNECTION',

  REFRESH_GAMES = 'REFRESH_GAMES',
  ADD_POTENTIAL_GAMES = 'ADD_POTENTIAL_GAMES',
  ADD_PLAYABLE_GAMES = 'ADD_PLAYABLE_GAMES',
  EDIT_PLAYABLE_GAME = 'EDIT_PLAYABLE_GAME',
  REMOVE_PLAYABLE_GAME = 'REMOVE_PLAYABLE_GAME',
  LAUNCH_GAME = 'LAUNCH_GAME',
  STOP_GAME = 'STOP_GAME',
  SELECT_GAME = 'SELECT_GAME',
  SET_POTENTIAL_GAME_TO_ADD = 'SET_POTENTIAL_GAME_TO_ADD',
  SET_GAME_TO_EDIT = 'SET_GAME_TO_EDIT',
  SORT_GAMES = 'SORT_GAMES',

  OPEN_GAME_ADD_MODAL = 'OPEN_GAME_ADD_MODAL',
  CLOSE_GAME_ADD_MODAL = 'CLOSE_GAME_ADD_MODAL',
  OPEN_IGDB_RESEARCH_MODAL = 'OPEN_IGDB_RESEARCH_MODAL',
  CLOSE_IGDB_RESEARCH_MODAL = 'CLOSE_IGDB_RESEARCH_MODAL',
  OPEN_TIME_PLAYED_EDITION_MODAL = 'OPEN_TIME_PLAYED_EDITION_MODAL',
  CLOSE_TIME_PLAYED_EDITION_MODAL = 'CLOSE_TIME_PLAYED_EDITION_MODAL',
  OPEN_POTENTIAL_GAMES_ADD_MODAL = 'OPEN_POTENTIAL_GAMES_ADD_MODAL',
  CLOSE_POTENTIAL_GAMES_ADD_MODAL = 'CLOSE_POTENTIAL_GAMES_ADD_MODAL',
  OPEN_SETTINGS_MODAL = 'OPEN_SETTINGS_MODAL',
  CLOSE_SETTINGS_MODAL = 'CLOSE_SETTINGS_MODAL'
}